Comprehensive analysis of Aurelium Skills for null pointer exceptions. Fixed several NPE bugs.
There are still lots of instances of unsafe access where assumptions have been made using string literals; TextUtil methods, which both accepts and returns null values; Locale, the resulting default language, and translations, ActiveMenu properties, and possibly others.

Assertions have been added for all instances of unsafe access, where certain behavior is assumed. Search of 'assert (null !=' for all instances of unsafe behavior that should be refactored.

This commit makes no changes in functional behavior to the plugin, and is meant to address potential null access violations, as well as to assert on undefined behavior. This is also a precursor to a full null analysis annotations commit of the source tree. Because the complete annotation will modify a lot of files I decided to split the pull request into two parts; one for the patches which address the problems found through analysis (without the annotations), and a second pull request with the annotations added. This also helps make it easier to debug any regressions in the event that I made an error.