Configurable mana ability item durability
When configured set the item durability when using mana abilities like Terraform and Treecapitator.

Includes the option to break items when they've reached (or go over) the max damage value. With the option for durability as max_blocks limiter This allows tools with less remaining durability to break more blocks as they're able to if the max_blocks is higher.