Add options for Treecapitator and Terraform to consume durability (#415)
When configured set the item durability when using mana abilities like Terraform and Treecapitator.
This includes the option to break items when they've reached (or go over) the max damage value.
The default behavior allows tools with less remaining durability to break more blocks as they're able to
if the max_blocks is higher.