Added blockTrigger collect
This allows blocks like beehives/-nests and cauldrons to be used as a source. When trigger is set to "collect" players can use shears (beehive/-nest) or bottles (all three) or buckets (cauldrons) to collect resources and get rewarded.

Interact would otherwise reward for any click made, instead of when resources are collected.